requirements.txt
streamlit
openai
python-docx
requests
beautifulsoup4
feedparser
pytz
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2